Add: 게시글 수정 삭제 로그인체크 추가

+++ front/src/community/CommunityPostDetail.tsx
@@ -102,8 +102,19 @@ const CommunityPostDetail = () => {
     }
   };
 
+  const checkAuthor = (author: string) => {
+    const userId = JSON.parse(user).id;
+    if (userId === author) {
+      return true;
+    }
+    return false;
+  };
+
   // 게시글 수정 이벤트
   const handleEditClick = (e: React.MouseEvent<HTMLButtonElement>) => {
+    if (!checkAuthor(fetchData?.author)) {
+      return alert("글쓴이만 수정 가능합니다.");
+    }
     e.preventDefault();
     navigate(`/community/write?id=${postId}`);
   };
@@ -111,6 +122,9 @@ const CommunityPostDetail = () => {
   const handleDeleteClick = async (e: React.MouseEvent<HTMLButtonElement>) => {
     e.preventDefault();
     try {
+      if (!checkAuthor(fetchData?.author)) {
+        return alert("글쓴이만 삭제 가능합니다.");
+      }
       if (window.confirm("정말로 삭제하시겠습니까?")) {
         await Api.delete(`posts/${postId}`);
         navigate("/community");
